refactor(town): extract town config helpers to town/config.js and update town_gen.js imports
